requirements.txt
python-telegram-bot==20.6
fastapi
uvicorn
python-dotenv
openai
nest_asyncio
